feat: Add DevTools component.
You can adjust the slider bar for each experiment to change whether you are in cohort or not. Since
'storage' event listener is not fired if localStorage.setItem() is called in the same tab, need to
look into an alternative solution. As a result, in order to see the changed settings, you need to
refresh the page at the moment.